Fix FPE & DLO on x86, .h cleanup, dbg info, -warni
- Corrects previous commit mistakes (b instead of kb, signatures of fuse
hooks, addresses in fprintf).
- Change debugging prefix to something more suitable and
review of broken english outputs.
- Reorganize headers (needed to fix compilation warning on
cloudfs_statfs).
- Remove unused variables.
- Harden cloudfs_file_size and make consistent use of it (replaced ftell
in cloudfs_object_read_fp)
- Changed static string in strptime call to empty string.
- Changed sleep inital time from 8 to 1 ()

* Issue 32 : Fix Floating point exception on x86 (use of off_t).
Also fixes incorrect uploads (1 or 3 GB instead of 10GB+)
* Issue 24 : Creates hook to .utimens (workaround only for the time
* being)
* Issue 68 : reduce problem by reducing the sleep initial value.
The problem on rm is identified and probably caused by issue 54 (when
removing a segment, we try to read a manifest ?
It triggers a 404 anyway, which triggers the call to sleep).
It will be easier to investigate thanks to added debug

Additional notes:
- Possibly identified issues:
When overwriting, the manifest is not recreated ?

Before copy, shouldn't we test the available disk free on the partition
containing temp_dir and fail early if insufficient space?

After copy, instead of keeping temp_file in temp_dir, shouldn't we
create
a symlink to src if still present, delete temp_file otherwise ?

- note on sanity check of segment_size and segment_above:
With very small value, we could crash the process because it would
create too many threads, hence the validity of the sanity check.

- note on tests:
did copy files of 14G and 70G to hubic.
did confirm file integrity with md5sum on the original and uploaded then
downloaded version of the original.

==-- Sorry for such a large commit --==